Fix System.Net.Primitives tests
Clean up System.Net.Primitives unit and functional tests. Moved some
functional tests to the unit tests projects since they are testing
non-public methods.

Change the `public` attribute on some of the CookieContainer` methods to
match the public ref contract. This wasn't done properly when this code
was first ported from .NET Framework to .NET Core. Some of the public
methods are only public in Desktop and not Core.

With these changes, the System.Net.Primitives tests can now build and
run with `/p:BuildTestsAgainstPackages=true`.

Fixes #7906.
